+++ app.js
@@ -4,8 +4,7 @@ import express from "express";
 // TODO: as we write router files, add import  for each:
 import testRecordRoutes from "./routes/testRecords.js"; //FYI we could call this (testRecordRoutes) anything suitable - it's an alias for router defined inside routes/testRecords.js
 
-//TEST - removing this to see if i can get heroku deployment working again
-//import appUserRoutes from "./routes/appUsers.js";
+import appUserRoutes from "./routes/appUsersX.js";
 
 import eventRoutes from "./routes/events.js";
 
@@ -29,8 +28,7 @@ app.use(
 app.use("/testrecords", testRecordRoutes);
 app.use("/events", eventRoutes);
 
-//TEST - removing this to see if i can get heroku deployment working again
-//app.use("/appusers", appUserRoutes);
+app.use("/appusers", appUserRoutes);
 
 export default app;
 // FYI the database PORT is now covered in bin/www.js
